Move PanelList to its own component file
