+++ frontend/src/App.jsx
@@ -1,39 +1,7 @@
-import { useState, useEffect } from 'react'
 import { Routes, Route } from 'react-router-dom'
 
-import panelService from './services/panels'
-
+import PanelList from './routes/PanelList'
 import PanelView from './routes/PanelView'
-
-import Thumbnail from './components/Thumbnail'
-import Panel from './components/Panel'
-
-const PanelList = () => {
-  const [panels, setPanels] = useState([])
-
-  useEffect(() => {
-    panelService
-      .getAll()
-      .then(initialPanels => {
-        setPanels(initialPanels)
-      })
-  }, [])
-
-  return (
-    <>
-      {panels.map(panel =>
-        <Thumbnail
-          key={panel.id}
-          width='150'
-          height='150'
-          title={panel.title}
-          href={`/panel/${panel.id}`}>
-          <Panel panel={panel} />
-        </Thumbnail>
-      )}
-    </>
-  )
-}
 
 const App = () => {
   return (
